backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
aiosqlite==0.20.0
pydantic==2.9.2
httpx==0.27.2
python-multipart==0.0.12
websockets==13.1
greenlet>=3.1.1
allure-python-commons==2.13.5
selenium>=4.15.0
webdriver-manager>=4.0.0
Appium-Python-Client>=3.0.0
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0
openpyxl>=3.1.0
openai>=1.0.0
